+++ swift/llm/utils/argument.py
@@ -1221,7 +1221,6 @@ class SftArguments(ArgumentsBase):
             metric_for_best_model='rouge-l' if self.predict_with_generate else 'loss',
             greater_is_better=self.predict_with_generate,
             full_determinism=self.full_determinism,
-            sortish_sampler=True,
             optim=self.optim,
             adam_beta1=self.adam_beta1,
             adam_beta2=self.adam_beta2,
@@ -1235,7 +1234,6 @@ class SftArguments(ArgumentsBase):
             ignore_data_skip=self.ignore_data_skip,
             ddp_backend=self.ddp_backend,
             gradient_checkpointing=self.gradient_checkpointing,
-            predict_with_generate=self.predict_with_generate,
             local_rank=self.local_rank,
             save_only_model=self.save_only_model,
             train_sampler_random=self.train_sampler_random,
