+++ swift/trainers/orpo_trainer.py
@@ -53,30 +53,7 @@ class ORPOTrainer(PushToMsHubMixin, SwiftMixin, HFORPOTrainer):
         kwargs['super_class'] = Trainer
         SwiftMixin.__init__(self, model, args, **kwargs)
         self._stored_metrics = defaultdict(lambda: defaultdict(list))
-        if not self.streaming and not self.lazy_tokenize:
-            train_ds_info = self.stat_dataset(self.train_dataset, self.is_encoder_decoder)
-
-            if self.eval_dataset is not None:
-                val_ds_info = self.stat_dataset(self.eval_dataset, self.is_encoder_decoder)
-                self.dataset_info = {'train_dataset': train_ds_info, 'val_dataset': val_ds_info}
-            else:
-                self.dataset_info = {'train_dataset': train_ds_info}
-        else:
-            self.dataset_info = {}
-        # performance
-        self.perf: Dict[str, Any] = {
-            'gen_time': 0.,
-            'gen_len': 0,
-            'memory': {},
-            'model': self.model.get_trainable_parameters() if hasattr(self.model, 'get_trainable_parameters') else None,
-        }
         self.model.config.model_type = self.model.config.model_type[:-1]  # remove suffix
-
-    def train(self, *args, **kwargs) -> torch.Tensor:
-        res = super().train(*args, **kwargs)
-        for i in range(torch.cuda.device_count()):
-            self.perf['memory'][f'cuda:{i}'] = f'{torch.cuda.max_memory_reserved(i)/1024/1024/1024:.2f}GiB'
-        return res
 
     def concatenated_forward(
         self, model: nn.Module, batch: Dict[str, Union[List, torch.LongTensor]]
